[BE] refactor: 로컬 이미지 처리 코드 클래스 분리
